correction des float en double

+++ WineBiblio/Business/Invoice.cs
@@ -9,8 +9,8 @@ namespace WineBiblio.Business
     public class Invoice
     {
         public int id_invoice { get; set; }
-        public float TVA { get; set; }
-        public float price_TTC { get; set; }
+        public double TVA { get; set; }
+        public double price_TTC { get; set; }
         public int id_order { get; set; }
         public int id_address { get; set; }
     }
